Added help + stylistic cleanups.
Added a -h option with help.

No need to generate a file in cases where it's never used.
Shouldn't generate files with fixed names into the current
working directory, temporarily or not. Remove file with a
trap (although not especially important now that it's not
generated in the current directory) so it can't evade removal
under early termination.